+++ core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `chg` **Restructure containers to support running other services alongside servers.**
+
+    *Related links:*
+    - [Pull Request #543][pr-543]
+
   * `add` **Introduce an async container strategy.**
 
     *Related links:*
@@ -775,6 +780,7 @@
     *Related links:*
     - [Pull Request #338][pr-338]
 
+[pr-543]: https://github.com/pakyow/pakyow/pull/543
 [pr-542]: https://github.com/pakyow/pakyow/pull/542
 [pr-541]: https://github.com/pakyow/pakyow/pull/541
 [pr-540]: https://github.com/pakyow/pakyow/pull/540